requirements/base.txt
Jinja2==2.11.2
python-dotenv==0.15.0
requests==2.25.1